massa_execution_worker::datastore_scan

Function merge_set_case_keys

fn merge_set_case_keys(
    absolute_set_datastore: Option<&Datastore>,
    key_updates: &BTreeMap<Vec<u8>, bool>,
    key_range: Option<&(Bound<Vec<u8>>, Bound<Vec<u8>>)>,
    count: Option<u32>,
) -> BTreeSet<Vec<u8>>

Builds the bounded result for a Set reset by streaming the absolute datastore in key order and overlaying the newer-layer updates, stopping as soon as count surviving keys are collected. A naive truncation of the absolute range would be wrong: a newer update may delete one of the first count keys, which must then be replaced by the next key.