massa_execution_worker

Function start_execution_worker

pub fn start_execution_worker(
    config: ExecutionConfig,
    final_state: Arc<RwLock<dyn FinalStateController>>,
    selector: Box<dyn SelectorController>,
    mip_store: MipStore,
    channels: ExecutionChannels,
    wallet: Arc<RwLock<Wallet>>,
    massa_metrics: MassaMetrics,
    event_cache: Box<dyn EventCacheController>,
) -> (Box<dyn ExecutionManager>, Box<dyn ExecutionController>)

Launches an execution worker thread and returns an ExecutionManager to interact with it

parameters

config: execution configuration

final_state: a thread-safe shared access to the final state for reading and writing

Returns

A pair (execution_manager, execution_controller) where:

execution_manager: allows to stop the worker

execution_controller: allows sending requests and notifications to the worker