massa_execution_worker

Module interface_impl

Implementation of the interface between massa-execution-worker and massa-sc-runtime. This allows the VM runtime to access the Massa execution context, for example to interact with the ledger. See the definition of Interface in the massa-sc-runtime crate for functional details.

Macros

context_guard 🔒

helper for locking the context mutex

Structs

InterfaceImpl

an implementation of the Interface trait (see massa-sc-runtime crate)

Functions

amount_from_native_amount 🔒

Helper function that creates an amount from a NativeAmount

amount_to_native_amount 🔒

Helper function that creates a NativeAmount from the amount internal representation

get_address_from_opt_or_context 🔒

Helper function to get the address from the option given as argument to some ABIs Fallback to the current context address if not provided.

massa_time_from_native_time 🔒

Helper function that creates an MassaTime from a NativeTime

massa_time_to_native_time 🔒

Helper function that creates a NativeTime from the MassaTime internal representation