massa_execution_worker::storage_backend

Trait StorageBackend

pub trait StorageBackend: Send + Sync {
    // Required methods
    fn write(&mut self, slot: &Slot, value: &[u8]);
    fn read(&self, slot: &Slot) -> Option<Vec<u8>>;
}

A trait that defines the interface for a storage backend for the dump-block feature.

Required Methods

fn write(&mut self, slot: &Slot, value: &[u8])

Writes the given value to the storage backend. The slot is used as the key to the value.

fn read(&self, slot: &Slot) -> Option<Vec<u8>>

Reads the value from the storage backend. The slot is used as the key to the value.